@@ -3,17 +3,29 @@ const themes = {
   // The default theme.
   default: {
     // React node added to the head of each document. Use this to import CSS, fonts, or whatever else you need.
-    head: <>
-      <link rel="preconnect" href="https://fonts.googleapis.com" />
-      <link rel="preconnect" href="https://fonts.gstatic.com" crossOrigin="true" />
-      {/* <link rel="stylesheet" href="https:use.typekit.net/ush4rsn.css" /> { /\* test Typekit *\/} */}
-      <link href="https://fonts.googleapis.com/css2?family=Source+Code+Pro&family=Source+Sans+Pro:ital,wght@0,300;0,400;0,700;1,400;1,700&display=swap" rel="stylesheet" />
-    </>,
-    site_title: 'Constellate',
-    site_logo: '/star.svg',
+    head: (
+      <>
+        <link rel="preconnect" href="https://fonts.googleapis.com" />
+        <link
+          rel="preconnect"
+          href="https://fonts.gstatic.com"
+          crossOrigin="true"
+        />
+        {/* <link rel="stylesheet" href="https:use.typekit.net/ush4rsn.css" /> { /\* test Typekit *\/} */}
+        <link
+          href="https://fonts.googleapis.com/css2?family=Source+Sans+3:ital,wght@0,200..900;1,200..900&family=Fira+Code:wght@300..700&display=swap"
+          rel="stylesheet"
+        />
+        <style>
+          {`@import url('/CascadiaCode.woff2'); @import url('/CascadiaCodeItalic.woff2');`}
+        </style>
+      </>
+    ),
+    site_title: "Constellate",
+    site_logo: "/star.svg",
     global: {
-      '#essay-title': {
-        fontSize: '1.563rem',
+      "#essay-title": {
+        fontSize: "1.563rem",
         fontWeight: 700,
       },
     },
@@ -21,53 +33,54 @@ const themes = {
     eui: {
       colors: {
         LIGHT: {
-          primary: '#215DB0',
-          accent: '#007067',
+          primary: "#215DB0",
+          accent: "#007067",
         },
         DARK: {
-          primary: '#4C90F0',
-          accent: '#13C9BA',
+          primary: "#4C90F0",
+          accent: "#13C9BA",
         },
       },
       font: {
-        family: "'Source Sans Pro', sans-serif",
-        familyCode: "'Source Code Pro', monospace",
+        family: "'Source Sans 3`, sans-serif",
+        familyCode: "'Cascadia Code', 'Fira Code', monospace",
       },
-    }
+    },
   },
 
   rho: {
-    head: <>
-      <link rel="stylesheet" href="https://use.typekit.net/rkc4ciw.css" />
-      <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/@xz/fonts@1/serve/cascadia-code.min.css" />
-    </>,
+    head: (
+      <>
+        <link rel="stylesheet" href="https://use.typekit.net/rkc4ciw.css" />
+      </>
+    ),
     site_title: "Pollard's Rho",
 
-    site_logo: '/pollardsrho.svg',
+    site_logo: "/pollardsrho.svg",
     global: {
-      '#essay-title': {
-        fontSize: '1.563rem',
+      "#essay-title": {
+        fontSize: "1.563rem",
         fontWeight: 700,
       },
     },
     eui: {
       colors: {
         LIGHT: {
-          primary: '#634DBF',
-          accent: '#7C327C',
+          primary: "#634DBF",
+          accent: "#7C327C",
         },
         DARK: {
-          primary: '#9881F3',
-          accent: '#BD6BBD',
+          primary: "#9881F3",
+          accent: "#BD6BBD",
         },
       },
       font: {
         family: "effra-cc, sans-serif",
         familyCode: "'Cascadia Code', monospace",
-        featureSettings: "'liga' 1, 'kern' 1, 'pnum' 1"
+        featureSettings: "'liga' 1, 'kern' 1, 'pnum' 1",
       },
-    }
-  }
+    },
+  },
 };
 
 export default themes;
