@@ -1,5 +1,5 @@
-import { transparentize, UseEuiTheme, useEuiFontSize } from '@elastic/eui';
-import { css } from '@emotion/react';
+import { transparentize, UseEuiTheme, useEuiFontSize } from "@elastic/eui";
+import { css } from "@emotion/react";
 
 // prettier-ignore
 export default function indexStyles({ euiTheme, colorMode }) {
